framework: tests: fix potential buffer overflow
When reading the buffer from the test process, we have a buffer memset
to 0 to ensure that the buffer will be terminated by a null character as
this will be printed as a string via printf().
We need to keep the last character for a null character as the buffer read
might not have one at the end.

Signed-off-by: Gabriel Huau <ghuau@witekio.us>