framework: report: fix unwanted angle bracket in the middle of line
If the buffer reported by the child's process is too big or small, the
buffer will be flushed in the middle of a string and at the next report
we will print another double angle bracket. We could have an output
like:

>> Usage: ping [-aAbBdDfhLnOqrR>> UvV64] [>> -c count] [-i >> inter>>
val] [-I interface]

As the angle brackets are part of a style for the test report, it should
be the responsibility of the test to define when and where it should be
printed.

This patch introduces a function name cobaye_run_command() that
encapsulated the system() C function with the angle brackets before.

Signed-off-by: Gabriel Huau <ghuau@witekio.us>